requirements.txt
alabaster==0.7.12
anyio==3.6.1
APScheduler==3.9.1
Babel==2.10.3
backports.zoneinfo==0.2.1
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
docutils==0.19
Foundations==2.0.7
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
imagesize==1.4.1
importlib-metadata==4.12.0
Jinja2==3.1.2
linecache2==1.0.0
Manager==2.0.3
MarkupSafe==2.1.1
ordereddict==1.1
packaging==21.3
peewee==3.15.1
Pillow==9.2.0
pkg_resources==0.0.0
Pygments==2.13.0
pyparsing==3.0.9
python-dotenv==0.21.0
python-telegram-bot==20.0a2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
qrcode==7.3.1
requests==2.28.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
snowballstemmer==2.2.0
socksio==1.0.0
Sphinx==5.1.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
tornado==6.2
traceback2==1.4.0
tzdata==2022.1
tzlocal==4.2
unittest2==1.1.0
urllib3==1.26.11
zipp==3.8.1
